paper_trading_system/requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
python-telegram-bot>=20.0
requests>=2.31.0
aiohttp>=3.9.0
